refactor redux to mainPage, clean up old panel separator
