+++ packages/client/src/pages/Main/containers/StatementsListBox/StatementListBox.tsx
@@ -18,7 +18,7 @@ import { useResizeObserver, useSearchParams } from "hooks";
 import React, { useEffect, useMemo, useState } from "react";
 import { BsInfoCircle } from "react-icons/bs";
 import { toast } from "react-toastify";
-import { setStatementListOpened } from "redux/features/layout/statementListOpenedSlice";
+import { setStatementListOpened } from "redux/features/layout/mainPage/statementListOpenedSlice";
 import { setShowWarnings } from "redux/features/statementEditor/showWarningsSlice";
 import { setDisableStatementListScroll } from "redux/features/statementList/disableStatementListScrollSlice";
 import { setRowsExpanded } from "redux/features/statementList/rowsExpandedSlice";
@@ -53,7 +53,7 @@ export const StatementListBox: React.FC = () => {
     (state) => state.statementList.rowsExpanded
   );
   const statementListOpened: boolean = useAppSelector(
-    (state) => state.layout.statementListOpened
+    (state) => state.layout.mainPage.statementListOpened
   );
   const isLoading: boolean = useAppSelector(
     (state) => state.statementList.isLoading
