+++ packages/client/src/redux/features/layout/mainPage/detailBoxMinimizedSlice.tsx
