+++ packages/client/src/redux/features/layout/mainPage/firstPanelExpandedSlice.tsx
