+++ packages/client/src/redux/features/layout/mainPage/thirdPanelExpandedSlice.tsx
